perf: 修改 ldap task 位置
